pyproject.toml
[tool.poetry]
name = "aaop"
version = "0.1.0"
description = "Captain logAR — AI-powered OTT platform operations"
authors = ["Fatih Ayaz"]
readme = "CHANGELOG.md"
packages = [
    { include = "shared" },
    { include = "backend" },
    { include = "apps" },
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = ">=0.115"
uvicorn = ">=0.32"
pydantic = ">=2.10"
pydantic-settings = ">=2.13"
structlog = ">=24.4"
tenacity = ">=9.1"
python-jose = { version = ">=3.5", extras = ["cryptography"] }
python-multipart = ">=0.0.18"
python-socketio = ">=5.16"
langgraph = ">=1.1"
langchain-anthropic = ">=1.4"
anthropic = ">=0.86"
redis = ">=7.3"
duckdb = ">=1.5"
aiosqlite = ">=0.22"
chromadb = ">=1.5"
boto3 = ">=1.34"
apscheduler = ">=3.10"
python-docx = ">=1.1"
plotly = ">=5.24"
kaleido = "==0.2.1"
bcrypt = ">=5.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0"
pytest-asyncio = ">=1.3"
pytest-cov = ">=7.0"
pytest-mock = ">=3.15"
ruff = ">=0.15"
mypy = ">=1.19"
pre-commit = ">=4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ── ruff ──────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 120
src = ["shared", "backend", "apps", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["shared", "backend", "apps"]

# ── mypy ──────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict = false

# ── pytest ────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
